fix: adjust the code position
